requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
tortoise-orm==0.20.0
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
apscheduler==3.10.4
aerich==0.7.2
jinja2==3.1.2
email-validator==2.1.0


pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
